+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Sidebar from "./components/Sidebar";
+import Sidebar from "./components/Aside/Sidebar";
 import Content from "./components/Contents";
 import styled from "styled-components";
 import { Routes, Route, Link } from "react-router-dom";
